Get SDGS indicators and values for a specific country

get indicators for a specific country and indicator

get_indicator(Country, indicator)

Arguments

Country

the M49 code for the country you are interested in

indicator

the code for the indicator you are interested in

Value

dataframe of indicator and values

Examples

require(SDGsR)
get_indicator(Country="12", indicator="15.6.1")
#>    goal target indicator          series
#> 1    15   15.6    15.6.1     ER_CBD_SMTA
#> 2    15   15.6    15.6.1     ER_CBD_SMTA
#> 3    15   15.6    15.6.1     ER_CBD_SMTA
#> 4    15   15.6    15.6.1     ER_CBD_SMTA
#> 5    15   15.6    15.6.1     ER_CBD_SMTA
#> 6    15   15.6    15.6.1     ER_CBD_SMTA
#> 7    15   15.6    15.6.1     ER_CBD_SMTA
#> 8    15   15.6    15.6.1     ER_CBD_SMTA
#> 9    15   15.6    15.6.1     ER_CBD_SMTA
#> 10   15   15.6    15.6.1     ER_CBD_SMTA
#> 11   15   15.6    15.6.1     ER_CBD_SMTA
#> 12   15   15.6    15.6.1     ER_CBD_SMTA
#> 13   15   15.6    15.6.1   ER_CBD_NAGOYA
#> 14   15   15.6    15.6.1   ER_CBD_NAGOYA
#> 15   15   15.6    15.6.1   ER_CBD_NAGOYA
#> 16   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 17   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 18   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 19   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 20   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 21   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 22   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 23   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 24   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 25   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#>                                                                                                                                                                                                                                         seriesDescription
#> 1                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 2                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 3                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 4                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 5                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 6                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 7                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 8                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 9                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 10                                                                                           Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 11                                                                                           Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 12                                                                                           Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 13                                                                                                                                                                                    Countries that are parties to the Nagoya Protocol (1 = YES; 0 = NO)
#> 14                                                                                                                                                                                    Countries that are parties to the Nagoya Protocol (1 = YES; 0 = NO)
#> 15                                                                                                                                                                                    Countries that are parties to the Nagoya Protocol (1 = YES; 0 = NO)
#> 16                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 17                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 18                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 19                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 20 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 21 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 22 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 23 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 24 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 25 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#>    seriesCount geoAreaCode geoAreaName timePeriodStart value valueType
#> 1         3492          12     Algeria            2012   365     Float
#> 2         3492          12     Algeria            2013   426     Float
#> 3         3492          12     Algeria            2014   473     Float
#> 4         3492          12     Algeria            2015   517     Float
#> 5         3492          12     Algeria            2016   536     Float
#> 6         3492          12     Algeria            2017   549     Float
#> 7         3492          12     Algeria            2018   561     Float
#> 8         3492          12     Algeria            2019   576     Float
#> 9         3492          12     Algeria            2020   588     Float
#> 10        3492          12     Algeria            2021   595     Float
#> 11        3492          12     Algeria            2022   609     Float
#> 12        3492          12     Algeria            2023   613     Float
#> 13         779          12     Algeria            2021     0     Float
#> 14         779          12     Algeria            2022     0     Float
#> 15         779          12     Algeria            2023     0     Float
#> 16         847          12     Algeria            2019     1     Float
#> 17         847          12     Algeria            2021     1     Float
#> 18         847          12     Algeria            2022     1     Float
#> 19         847          12     Algeria            2023     1     Float
#> 20        2159          12     Algeria            2016     0     Float
#> 21        2159          12     Algeria            2017     0     Float
#> 22        2159          12     Algeria            2018     0     Float
#> 23        2159          12     Algeria            2019     0     Float
#> 24        2159          12     Algeria            2020     0     Float
#> 25        2159          12     Algeria            2021     0     Float
#>    time_detail timeCoverage upperBound lowerBound basePeriod
#> 1           NA           NA         NA         NA         NA
#> 2           NA           NA         NA         NA         NA
#> 3           NA           NA         NA         NA         NA
#> 4           NA           NA         NA         NA         NA
#> 5           NA           NA         NA         NA         NA
#> 6           NA           NA         NA         NA         NA
#> 7           NA           NA         NA         NA         NA
#> 8           NA           NA         NA         NA         NA
#> 9           NA           NA         NA         NA         NA
#> 10          NA           NA         NA         NA         NA
#> 11          NA           NA         NA         NA         NA
#> 12          NA           NA         NA         NA         NA
#> 13          NA           NA         NA         NA         NA
#> 14          NA           NA         NA         NA         NA
#> 15          NA           NA         NA         NA         NA
#> 16          NA           NA         NA         NA         NA
#> 17          NA           NA         NA         NA         NA
#> 18          NA           NA         NA         NA         NA
#> 19          NA           NA         NA         NA         NA
#> 20          NA           NA         NA         NA         NA
#> 21          NA           NA         NA         NA         NA
#> 22          NA           NA         NA         NA         NA
#> 23          NA           NA         NA         NA         NA
#> 24          NA           NA         NA         NA         NA
#> 25          NA           NA         NA         NA         NA
#>                                                                                                                                                                                                                                                             source
#> 1                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 2                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 3                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 4                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 5                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 6                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 7                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 8                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 9                                                                                                                                                                                                                                             ITPGRFA Secretariat 
#> 10                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 11                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 12                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 13 https://www.cbd.int/abs/nagoya-protocol/signatories/default.shtml, Access and Benefit-sharing Clearing-house https://absch.cbd.int/, http://www.fao.org/plant-treaty/countries/membership/en/,  The Online Reporting System on Compliance of the International 
#> 14 https://www.cbd.int/abs/nagoya-protocol/signatories/default.shtml, Access and Benefit-sharing Clearing-house https://absch.cbd.int/, http://www.fao.org/plant-treaty/countries/membership/en/,  The Online Reporting System on Compliance of the International 
#> 15 https://www.cbd.int/abs/nagoya-protocol/signatories/default.shtml, Access and Benefit-sharing Clearing-house https://absch.cbd.int/, http://www.fao.org/plant-treaty/countries/membership/en/,  The Online Reporting System on Compliance of the International 
#> 16                                                                                                                                                                                                                                 https://absch.cbd.int/countries
#> 17                                                                                                                                                                                                                                 https://absch.cbd.int/countries
#> 18                                                                                                                                                                                                                                 https://absch.cbd.int/countries
#> 19                                                                                                                                                                                                                                 https://absch.cbd.int/countries
#> 20                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 21                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 22                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 23                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 24                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#> 25                                                                                                                                                                                                                                            ITPGRFA Secretariat 
#>    geoInfoUrl
#> 1          NA
#> 2          NA
#> 3          NA
#> 4          NA
#> 5          NA
#> 6          NA
#> 7          NA
#> 8          NA
#> 9          NA
#> 10         NA
#> 11         NA
#> 12         NA
#> 13         NA
#> 14         NA
#> 15         NA
#> 16         NA
#> 17         NA
#> 18         NA
#> 19         NA
#> 20         NA
#> 21         NA
#> 22         NA
#> 23         NA
#> 24         NA
#> 25         NA
#>                                                                                                                                                                                                                                                          footnotes
#> 1                                                                                                                                                                                                                                                Cumulative value.
#> 2                                                                                                                                                                                                                                                Cumulative value.
#> 3                                                                                                                                                                                                                                                Cumulative value.
#> 4                                                                                                                                                                                                                                                Cumulative value.
#> 5                                                                                                                                                                                                                                                Cumulative value.
#> 6                                                                                                                                                                                                                                                Cumulative value.
#> 7                                                                                                                                                                                                                                                Cumulative value.
#> 8                                                                                                                                                                                                                                                Cumulative value.
#> 9                                                                                                                                                                                                                                                Cumulative value.
#> 10                                                                                                                                                                                                                                               Cumulative value.
#> 11                                                                                                                                                                                                                                               Cumulative value.
#> 12                                                                                                                                                                                                                                               Cumulative value.
#> 13 Parties  refers to Parties to the Convention on Biological Diversity that have ratified, accepted, approved or acceded the Nagoya Protocol, including those for which the Protocol has not yet entered into force following the 90 days period set out in Artic
#> 14 Parties  refers to Parties to the Convention on Biological Diversity that have ratified, accepted, approved or acceded the Nagoya Protocol, including those for which the Protocol has not yet entered into force following the 90 days period set out in Artic
#> 15 Parties  refers to Parties to the Convention on Biological Diversity that have ratified, accepted, approved or acceded the Nagoya Protocol, including those for which the Protocol has not yet entered into force following the 90 days period set out in Artic
#> 16                                                            ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 17                                                            ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 18                                                            ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 19                                                            ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 20                                                                                                                                                                                                                                                                
#> 21                                                                                                                                                                                                                                                                
#> 22                                                                                                                                                                                                                                                                
#> 23                                                                                                                                                                                                                                                                
#> 24                                                                                                                                                                                                                                                                
#> 25                                                                                                                                                                                                                                                                
#>    Nature Reporting Type
#> 1       C              G
#> 2       C              G
#> 3       C              G
#> 4       C              G
#> 5       C              G
#> 6       C              G
#> 7       C              G
#> 8       C              G
#> 9       C              G
#> 10      C              G
#> 11      C              G
#> 12      C              G
#> 13      C              G
#> 14      C              G
#> 15      C              G
#> 16      C              G
#> 17      C              G
#> 18      C              G
#> 19      C              G
#> 20      C              G
#> 21      C              G
#> 22      C              G
#> 23      C              G
#> 24      C              G
#> 25      C              G