get_indicator

get indicators for a specific country and indicator

get_indicator(Country = "578", indicator = "1.1.1")

Arguments

Country
the M49 code for the country you are interested in

indicator
the code for the indicator you are interested in

Value

dataframe of indicator and values

Examples

get_indicator(Country="12", indicator="15.6.1")

#>    goal target indicator          series
#> 1    15   15.6    15.6.1     ER_CBD_SMTA
#> 2    15   15.6    15.6.1     ER_CBD_SMTA
#> 3    15   15.6    15.6.1     ER_CBD_SMTA
#> 4    15   15.6    15.6.1     ER_CBD_SMTA
#> 5    15   15.6    15.6.1     ER_CBD_SMTA
#> 6    15   15.6    15.6.1     ER_CBD_SMTA
#> 7    15   15.6    15.6.1     ER_CBD_SMTA
#> 8    15   15.6    15.6.1     ER_CBD_SMTA
#> 9    15   15.6    15.6.1     ER_CBD_SMTA
#> 10   15   15.6    15.6.1     ER_CBD_SMTA
#> 11   15   15.6    15.6.1   ER_CBD_NAGOYA
#> 12   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 13   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 14   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 15   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 16   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 17   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 18   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 19   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 20   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 21   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 22   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 23   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 24   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 25   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#>                                                                                                                                                                                                                                         seriesDescription
#> 1                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 2                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 3                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 4                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 5                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 6                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 7                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 8                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 9                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 10                                                                                           Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 11                                                                                                                                                                                    Countries that are parties to the Nagoya Protocol (1 = YES; 0 = NO)
#> 12                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 13 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 14 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 15 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 16 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 17 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 18 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 19                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 20                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 21                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 22                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 23                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 24                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 25                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#>    seriesCount geoAreaCode geoAreaName timePeriodStart value valueType
#> 1         2874          12     Algeria            2012   365     Float
#> 2         2874          12     Algeria            2013   425     Float
#> 3         2874          12     Algeria            2014   472     Float
#> 4         2874          12     Algeria            2015   516     Float
#> 5         2874          12     Algeria            2016   532     Float
#> 6         2874          12     Algeria            2017   543     Float
#> 7         2874          12     Algeria            2018   547     Float
#> 8         2874          12     Algeria            2019   553     Float
#> 9         2874          12     Algeria            2020   553     Float
#> 10        2874          12     Algeria            2021   553     Float
#> 11         277          12     Algeria            2012     0     Float
#> 12         277          12     Algeria            2012     1     Float
#> 13        1600          12     Algeria            2016     0     Float
#> 14        1600          12     Algeria            2017     0     Float
#> 15        1600          12     Algeria            2018     0     Float
#> 16        1600          12     Algeria            2019     0     Float
#> 17        1600          12     Algeria            2020     0     Float
#> 18        1600          12     Algeria            2021     0     Float
#> 19        2679          12     Algeria            2012     1     Float
#> 20        2679          12     Algeria            2013     1     Float
#> 21        2679          12     Algeria            2014     1     Float
#> 22        2679          12     Algeria            2015     1     Float
#> 23        2679          12     Algeria            2016     1     Float
#> 24        2679          12     Algeria            2017     1     Float
#> 25        2679          12     Algeria            2018     1     Float
#>    time_detail timeCoverage upperBound lowerBound basePeriod
#> 1           NA           NA         NA         NA         NA
#> 2           NA           NA         NA         NA         NA
#> 3           NA           NA         NA         NA         NA
#> 4           NA           NA         NA         NA         NA
#> 5           NA           NA         NA         NA         NA
#> 6           NA           NA         NA         NA         NA
#> 7           NA           NA         NA         NA         NA
#> 8           NA           NA         NA         NA         NA
#> 9           NA           NA         NA         NA         NA
#> 10          NA           NA         NA         NA         NA
#> 11          NA           NA         NA         NA         NA
#> 12          NA           NA         NA         NA         NA
#> 13          NA           NA         NA         NA         NA
#> 14          NA           NA         NA         NA         NA
#> 15          NA           NA         NA         NA         NA
#> 16          NA           NA         NA         NA         NA
#> 17          NA           NA         NA         NA         NA
#> 18          NA           NA         NA         NA         NA
#> 19          NA           NA         NA         NA         NA
#> 20          NA           NA         NA         NA         NA
#> 21          NA           NA         NA         NA         NA
#> 22          NA           NA         NA         NA         NA
#> 23          NA           NA         NA         NA         NA
#> 24          NA           NA         NA         NA         NA
#> 25          NA           NA         NA         NA         NA
#>                                                                                                                                                                                                                                                                                                                                                                               source
#> 1                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 2                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 3                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 4                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 5                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 6                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 7                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 8                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 9                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 10                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 11 https://www.cbd.int/abs/nagoya-protocol/signatories/default.shtml, Access and Benefit-sharing Clearing-house https://absch.cbd.int/, http://www.fao.org/plant-treaty/countries/membership/en/,  The Online Reporting System on Compliance of the International Treaty on PGRFA http://faoitpgrfa.ort-production.linode.unep-wcmc.org and  Easy-SMTA https://mls.planttreaty.org ,
#> 12                                                                                                                                                                                                                                                                                                                                                   https://absch.cbd.int/countries
#> 13                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 14                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 15                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 16                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 17                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 18                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 19                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 20                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 21                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 22                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 23                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 24                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 25                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#>    geoInfoUrl
#> 1          NA
#> 2          NA
#> 3          NA
#> 4          NA
#> 5          NA
#> 6          NA
#> 7          NA
#> 8          NA
#> 9          NA
#> 10         NA
#> 11         NA
#> 12         NA
#> 13         NA
#> 14         NA
#> 15         NA
#> 16         NA
#> 17         NA
#> 18         NA
#> 19         NA
#> 20         NA
#> 21         NA
#> 22         NA
#> 23         NA
#> 24         NA
#> 25         NA
#>                                                                                                                                                                                                                                                                                                                                                      footnotes
#> 1                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 2                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 3                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 4                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 5                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 6                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 7                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 8                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 9                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 10                                                                                                                                                                                                                                                                                                                                           Cumulative value.
#> 11 Parties  refers to Parties to the Convention on Biological Diversity that have ratified, accepted, approved or acceded the Nagoya Protocol, including those for which the Protocol has not yet entered into force following the 90 days period set out in Article 33.2 of the Protocol. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 12                                                                                                                                                        ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 13                                                                                                                                                                                                                                                                                                                                                            
#> 14                                                                                                                                                                                                                                                                                                                                                            
#> 15                                                                                                                                                                                                                                                                                                                                                            
#> 16                                                                                                                                                                                                                                                                                                                                                            
#> 17                                                                                                                                                                                                                                                                                                                                                            
#> 18                                                                                                                                                                                                                                                                                                                                                            
#> 19                                                                                                                                                                                                                                                                                                                                                            
#> 20                                                                                                                                                                                                                                                                                                                                                            
#> 21                                                                                                                                                                                                                                                                                                                                                            
#> 22                                                                                                                                                                                                                                                                                                                                                            
#> 23                                                                                                                                                                                                                                                                                                                                                            
#> 24                                                                                                                                                                                                                                                                                                                                                            
#> 25                                                                                                                                                                                                                                                                                                                                                            
#>    attributes.Nature attributes.Units Reporting Type
#> 1                  C           NUMBER              G
#> 2                  C           NUMBER              G
#> 3                  C           NUMBER              G
#> 4                  C           NUMBER              G
#> 5                  C           NUMBER              G
#> 6                  C           NUMBER              G
#> 7                  C           NUMBER              G
#> 8                  C           NUMBER              G
#> 9                  C           NUMBER              G
#> 10                 C           NUMBER              G
#> 11                 C           NUMBER              G
#> 12                 C           NUMBER              G
#> 13                 C           NUMBER              G
#> 14                 C           NUMBER              G
#> 15                 C           NUMBER              G
#> 16                 C           NUMBER              G
#> 17                 C           NUMBER              G
#> 18                 C           NUMBER              G
#> 19                 C           NUMBER              G
#> 20                 C           NUMBER              G
#> 21                 C           NUMBER              G
#> 22                 C           NUMBER              G
#> 23                 C           NUMBER              G
#> 24                 C           NUMBER              G
#> 25                 C           NUMBER              G