get_indicator

get indicators for a specific country and indicator

get_indicator(Country = "578", indicator = "1.1.1")

Arguments

Country
the M49 code for the country you are interested in

indicator
the code for the indicator you are interested in

Value

dataframe of indicator and values

Examples

get_indicator(Country="12", indicator="15.6.1")

#>    goal target indicator          series
#> 1    15   15.6    15.6.1     ER_CBD_SMTA
#> 2    15   15.6    15.6.1     ER_CBD_SMTA
#> 3    15   15.6    15.6.1     ER_CBD_SMTA
#> 4    15   15.6    15.6.1     ER_CBD_SMTA
#> 5    15   15.6    15.6.1     ER_CBD_SMTA
#> 6    15   15.6    15.6.1     ER_CBD_SMTA
#> 7    15   15.6    15.6.1     ER_CBD_SMTA
#> 8    15   15.6    15.6.1     ER_CBD_SMTA
#> 9    15   15.6    15.6.1   ER_CBD_NAGOYA
#> 10   15   15.6    15.6.1 ER_CBD_ABSCLRHS
#> 11   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 12   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 13   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 14   15   15.6    15.6.1 ER_CBD_ORSPGRFA
#> 15   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 16   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 17   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 18   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 19   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 20   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 21   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#> 22   15   15.6    15.6.1 ER_CBD_PTYPGRFA
#>                                                                                                                                                                                                                                         seriesDescription
#> 1                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 2                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 3                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 4                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 5                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 6                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 7                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 8                                                                                            Total reported number of Standard Material Transfer Agreements (SMTAs) transferring plant genetic resources for food and agriculture to the country (number)
#> 9                                                                                                                                                                                     Countries that are parties to the Nagoya Protocol (1 = YES; 0 = NO)
#> 10                                                                                           Countries that have legislative, administrative and policy framework or measures reported to the Access and Benefit-Sharing Clearing-House (1 = YES; 0 = NO)
#> 11 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 12 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 13 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 14 Countries that have legislative, administrative and policy framework or measures reported through the Online Reporting System on Compliance  of the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 15                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 16                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 17                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 18                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 19                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 20                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 21                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#> 22                                                                                                       Countries that are contracting Parties to the International Treaty on Plant Genetic Resources for Food and Agriculture (PGRFA) (1 = YES; 0 = NO)
#>    seriesCount geoAreaCode geoAreaName timePeriodStart value valueType
#> 1         2304          12     Algeria            2012   369     Float
#> 2         2304          12     Algeria            2013   406     Float
#> 3         2304          12     Algeria            2014   434     Float
#> 4         2304          12     Algeria            2015   630     Float
#> 5         2304          12     Algeria            2016   646     Float
#> 6         2304          12     Algeria            2017   657     Float
#> 7         2304          12     Algeria            2018   661     Float
#> 8         2304          12     Algeria            2019   661     Float
#> 9          257          12     Algeria            2012     0     Float
#> 10         257          12     Algeria            2012     1     Float
#> 11        1072          12     Algeria            2016     0     Float
#> 12        1072          12     Algeria            2017     0     Float
#> 13        1072          12     Algeria            2018     0     Float
#> 14        1072          12     Algeria            2019     0     Float
#> 15        2144          12     Algeria            2012     1     Float
#> 16        2144          12     Algeria            2013     1     Float
#> 17        2144          12     Algeria            2014     1     Float
#> 18        2144          12     Algeria            2015     1     Float
#> 19        2144          12     Algeria            2016     1     Float
#> 20        2144          12     Algeria            2017     1     Float
#> 21        2144          12     Algeria            2018     1     Float
#> 22        2144          12     Algeria            2019     1     Float
#>    time_detail timeCoverage upperBound lowerBound basePeriod
#> 1           NA           NA         NA         NA         NA
#> 2           NA           NA         NA         NA         NA
#> 3           NA           NA         NA         NA         NA
#> 4           NA           NA         NA         NA         NA
#> 5           NA           NA         NA         NA         NA
#> 6           NA           NA         NA         NA         NA
#> 7           NA           NA         NA         NA         NA
#> 8           NA           NA         NA         NA         NA
#> 9           NA           NA         NA         NA         NA
#> 10          NA           NA         NA         NA         NA
#> 11          NA           NA         NA         NA         NA
#> 12          NA           NA         NA         NA         NA
#> 13          NA           NA         NA         NA         NA
#> 14          NA           NA         NA         NA         NA
#> 15          NA           NA         NA         NA         NA
#> 16          NA           NA         NA         NA         NA
#> 17          NA           NA         NA         NA         NA
#> 18          NA           NA         NA         NA         NA
#> 19          NA           NA         NA         NA         NA
#> 20          NA           NA         NA         NA         NA
#> 21          NA           NA         NA         NA         NA
#> 22          NA           NA         NA         NA         NA
#>                                                                                                                                                                                                                                                                                                                                                                               source
#> 1                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 2                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 3                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 4                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 5                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 6                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 7                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 8                                                                                                                                                                                                                                                                                                                                                               ITPGRFA Secretariat 
#> 9  https://www.cbd.int/abs/nagoya-protocol/signatories/default.shtml, Access and Benefit-sharing Clearing-house https://absch.cbd.int/, http://www.fao.org/plant-treaty/countries/membership/en/,  The Online Reporting System on Compliance of the International Treaty on PGRFA http://faoitpgrfa.ort-production.linode.unep-wcmc.org and  Easy-SMTA https://mls.planttreaty.org ,
#> 10                                                                                                                                                                                                                                                                                                                                                   https://absch.cbd.int/countries
#> 11                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 12                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 13                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 14                                                                                                                                                                                                                                                                                                                      Secretariat of the Convention on Biological Diversity (CBD) 
#> 15                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 16                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 17                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 18                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 19                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 20                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 21                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#> 22                                                                                                                                                                                                                                                                                                                                                              ITPGRFA Secretariat 
#>    geoInfoUrl
#> 1          NA
#> 2          NA
#> 3          NA
#> 4          NA
#> 5          NA
#> 6          NA
#> 7          NA
#> 8          NA
#> 9          NA
#> 10         NA
#> 11         NA
#> 12         NA
#> 13         NA
#> 14         NA
#> 15         NA
#> 16         NA
#> 17         NA
#> 18         NA
#> 19         NA
#> 20         NA
#> 21         NA
#> 22         NA
#>                                                                                                                                                                                                                                                                                                                                                      footnotes
#> 1                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 2                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 3                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 4                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 5                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 6                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 7                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 8                                                                                                                                                                                                                                                                                                                                            Cumulative value.
#> 9  Parties  refers to Parties to the Convention on Biological Diversity that have ratified, accepted, approved or acceded the Nagoya Protocol, including those for which the Protocol has not yet entered into force following the 90 days period set out in Article 33.2 of the Protocol. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 10                                                                                                                                                        ABS measures reported refers to legislative, administrative and policy framework or measures reported to the ABS Clearing-House. For values:  TRUE = 1.0; FALSE = 0.0. Data as per 15 February 2017.
#> 11                                                                                                                                                                                                                                                                                                                                                            
#> 12                                                                                                                                                                                                                                                                                                                                                            
#> 13                                                                                                                                                                                                                                                                                                                                                            
#> 14                                                                                                                                                                                                                                                                                                                                                            
#> 15                                                                                                                                                                                                                                                                                                                                                            
#> 16                                                                                                                                                                                                                                                                                                                                                            
#> 17                                                                                                                                                                                                                                                                                                                                                            
#> 18                                                                                                                                                                                                                                                                                                                                                            
#> 19                                                                                                                                                                                                                                                                                                                                                            
#> 20                                                                                                                                                                                                                                                                                                                                                            
#> 21                                                                                                                                                                                                                                                                                                                                                            
#> 22                                                                                                                                                                                                                                                                                                                                                            
#>    attributes.Nature attributes.Units Reporting Type
#> 1                 NA           NUMBER              G
#> 2                 NA           NUMBER              G
#> 3                 NA           NUMBER              G
#> 4                 NA           NUMBER              G
#> 5                 NA           NUMBER              G
#> 6                 NA           NUMBER              G
#> 7                 NA           NUMBER              G
#> 8                 NA           NUMBER              G
#> 9                 NA           NUMBER              G
#> 10                NA           NUMBER              G
#> 11                NA           NUMBER              G
#> 12                NA           NUMBER              G
#> 13                NA           NUMBER              G
#> 14                NA           NUMBER              G
#> 15                NA           NUMBER              G
#> 16                NA           NUMBER              G
#> 17                NA           NUMBER              G
#> 18                NA           NUMBER              G
#> 19                NA           NUMBER              G
#> 20                NA           NUMBER              G
#> 21                NA           NUMBER              G
#> 22                NA           NUMBER              G